requirements.txt
accelerate==0.24.1
boto3==1.33.2
botocore==1.33.2
certifi==2023.11.17
charset-normalizer==3.3.2
diffusers==0.23.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.8.1
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
urllib3
zipp==3.17.0